move friend methods to the correct controller
